Add proto files and update CmakeLists.txt
- Add proto files for Session, CameraModel, and Query
- Update CmakeLists.txt to reflect GRPC and protobuf support
- Use .gitignore to ignore protobuf and GRPC generated source files
- not working yet

+++ server/lib/data/Session.h
@@ -4,18 +4,18 @@ enum SessionType { HTTP_POST = 0, BOSSWAVE = 1 };
 
 class Session {
 public:
-  long id;
+  unsigned long long id;
   SessionType type;
 
 public:
-  long overallStart;
-  long overallEnd;
-  long featuresStart;
-  long featuresEnd;
-  long wordsStart;
-  long wordsEnd;
-  long signaturesStart;
-  long signaturesEnd;
-  long perspectiveStart;
-  long perspectiveEnd;
+  unsigned long long overallStart;
+  unsigned long long overallEnd;
+  unsigned long long featuresStart;
+  unsigned long long featuresEnd;
+  unsigned long long wordsStart;
+  unsigned long long wordsEnd;
+  unsigned long long signaturesStart;
+  unsigned long long signaturesEnd;
+  unsigned long long perspectiveStart;
+  unsigned long long perspectiveEnd;
 };
