+++ server/lib/util/Time.cpp
@@ -3,7 +3,7 @@
 #include <sys/time.h>
 
 /* Returns current time in milliseconds. */
-long getTime() {
+unsigned long long getTime() {
   struct timeval tv;
   gettimeofday(&tv, nullptr);
   return tv.tv_sec * 1000 + tv.tv_usec / 1000;
