+++ server/lib/util/Time.h
@@ -1,4 +1,4 @@
 #pragma once
 
 /* Returns current time in milliseconds. */
-long getTime();
+unsigned long long getTime();
